+++ models/demos/deepseek_v3_b1/weights/overlap/packing.py
@@ -2,125 +2,15 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+"""Overlap packing — fuse multiple tensors into a single L1 buffer with per-core shards."""
+
 from dataclasses import dataclass
 
 import numpy as np
 import torch
 
 import ttnn
-
-_DTYPE_ELEMENT_BYTES = {
-    ttnn.bfloat16: 2,
-    ttnn.float32: 4,
-    ttnn.uint16: 2,
-    ttnn.uint32: 4,
-    ttnn.int32: 4,
-}
-
-
-@dataclass(frozen=True)
-class OverlappedTensorSpec:
-    """Describes one sub-tensor within a fused raw-byte buffer.
-
-    Multiple ``OverlappedTensorSpec`` instances share a single L1
-    buffer whose per-core shard is zero-padded to a common maximum
-    byte size.  Each spec carries the core range, logical tensor shape,
-    dtype, and tile dimensions needed to pack its portion of the shard.
-
-    Tile byte sizes for BFP formats are computed from ``tile_h`` and
-    ``tile_w`` rather than stored as constants, so non-standard tile
-    shapes (e.g. 1x32, 16x32) are handled automatically.
-
-    Shape tuples follow (height, width) convention.
-
-    ``sharding`` controls how cores partition the per-device tensor:
-
-    - ``WIDTH_SHARDED``: each core gets a column slice (full height,
-      partial width).  ``num_cores`` divides the width.
-    - ``HEIGHT_SHARDED``: each core gets a row slice (partial height,
-      full width).  ``num_cores`` divides the height.
-
-    When data is preprocessed (shuffled / block-sharded) before
-    overlapping, ``raw_tensor_shape`` reflects the physical layout
-    while ``logical_tensor_shape`` preserves the original per-device
-    shape for downstream consumers.
-
-    ``name`` is an optional logical identifier used when the spec is
-    embedded in a :class:`RegionSpec` for cache fingerprinting.
-    """
-
-    core_range_set: ttnn.CoreRangeSet
-    raw_tensor_shape: tuple[int, int]
-    dtype: ttnn.DataType
-    sharding: ttnn.TensorMemoryLayout = ttnn.TensorMemoryLayout.WIDTH_SHARDED
-
-    tile_h: int = 32
-    tile_w: int = 32
-
-    tp_dim: tuple[int | None, int | None] = (None, None)
-
-    logical_tensor_shape: tuple[int, int] | None = None
-
-    name: str = ""
-
-    def _tile_bytes(self) -> int:
-        num_elements = self.tile_h * self.tile_w
-        if self.dtype in (ttnn.bfloat8_b, ttnn.bfloat4_b):
-            _L1_ALIGNMENT = 16
-            num_exponents = num_elements // 16
-            exponent_bytes = (num_exponents + _L1_ALIGNMENT - 1) // _L1_ALIGNMENT * _L1_ALIGNMENT
-            mantissa_bytes = num_elements if self.dtype == ttnn.bfloat8_b else num_elements // 2
-            return exponent_bytes + mantissa_bytes
-        return num_elements * _DTYPE_ELEMENT_BYTES[self.dtype]
-
-    def _dim_tp(self, tensor_dim: int, mesh_shape: tuple[int, int]) -> int:
-        """TP factor for a single tensor dimension (0=height, 1=width)."""
-        result = 1
-        for mesh_dim, d in enumerate(self.tp_dim):
-            if d == tensor_dim:
-                result *= mesh_shape[mesh_dim]
-        return result
-
-    def _dim_slice_idx(self, tensor_dim: int, row: int, col: int, mesh_shape: tuple[int, int]) -> int:
-        """Slice index for a single tensor dimension at mesh coordinate (row, col)."""
-        idx = 0
-        stride = 1
-        mesh_coord = (row, col)
-        for mesh_dim in reversed(range(2)):
-            if self.tp_dim[mesh_dim] == tensor_dim:
-                idx += mesh_coord[mesh_dim] * stride
-                stride *= mesh_shape[mesh_dim]
-        return idx
-
-    def tp(self, mesh_shape: tuple[int, int]) -> int:
-        """Total TP factor (product of all non-None mesh dimensions)."""
-        return self._dim_tp(0, mesh_shape) * self._dim_tp(1, mesh_shape)
-
-    def per_device_height(self, mesh_shape: tuple[int, int]) -> int:
-        return self.raw_tensor_shape[0] // self._dim_tp(0, mesh_shape)
-
-    def per_device_width(self, mesh_shape: tuple[int, int]) -> int:
-        return self.raw_tensor_shape[1] // self._dim_tp(1, mesh_shape)
-
-    def shard_shape(self, mesh_shape: tuple[int, int]) -> tuple[int, int]:
-        pdh = self.per_device_height(mesh_shape)
-        pdw = self.per_device_width(mesh_shape)
-        num_cores = self.core_range_set.num_cores()
-        if self.sharding == ttnn.TensorMemoryLayout.WIDTH_SHARDED:
-            return (pdh, pdw // num_cores)
-        else:
-            return (pdh // num_cores, pdw)
-
-    def tiles_per_shard(self, mesh_shape: tuple[int, int]) -> int:
-        sh, sw = self.shard_shape(mesh_shape)
-        return (sh // self.tile_h) * (sw // self.tile_w)
-
-    def shard_bytes(self, mesh_shape: tuple[int, int]) -> int:
-        return self.tiles_per_shard(mesh_shape) * self._tile_bytes()
-
-
-def max_shard_bytes(shard_specs: list[list[OverlappedTensorSpec]], mesh_shape: tuple[int, int]) -> int:
-    return max(sum(spec.shard_bytes(mesh_shape) for spec in lane) for lane in shard_specs)
+from models.demos.deepseek_v3_b1.weights.overlap.spec import OverlappedTensorSpec, max_shard_bytes
 
 
 def tilize_and_pack(data_2d: torch.Tensor, spec: OverlappedTensorSpec) -> bytes:
@@ -302,8 +192,6 @@ def overlap_tensors(
                 entry.spec.per_device_width(mesh_shape),
             )
             result[entry.name] = OverlappedTensor(
-                # fused tensor is just a reference to the underlying fused tensor, that
-                # stores all overlapped tensors.
                 fused_tensor=fused,
                 tensor_shape=ts,
                 shard_shape=entry.spec.shard_shape(mesh_shape),
@@ -318,22 +206,14 @@ def overlap_tensors(
 
 
 def tilize_and_pack_bfp8(data_2d: torch.Tensor, tile_h: int = 32, tile_w: int = 32) -> bytes:
-    """Tilize a 2-D tensor and pack as BFP8_b raw bytes.
-
-    Delegates to the C++ ``pack_as_bfp8_tiles`` via nanobind.
-    """
+    """Tilize a 2-D tensor and pack as BFP8_b raw bytes."""
     H, W = data_2d.shape
     data_np = data_2d.contiguous().float().numpy()
     return ttnn._ttnn.core.tilize_and_pack_bfp8_b(data_np, H, W, tile_h, tile_w)
 
 
 def tilize_and_pack_bfloat16(data_2d: torch.Tensor, tile_h: int = 32, tile_w: int = 32) -> bytes:
-    """Tilize a 2-D tensor and pack as bfloat16 (Float16_b) raw bytes.
-
-    Each tile is 2048 bytes: 1024 elements x 2 bytes, stored in
-    face order (face0, face1, face2, face3), row-major within each
-    face.  bfloat16 is the top 16 bits of IEEE-754 float32.
-    """
+    """Tilize a 2-D tensor and pack as bfloat16 (Float16_b) raw bytes."""
     H, W = data_2d.shape
     face_h, face_w = tile_h // 2, tile_w // 2
     tr, tc = H // tile_h, W // tile_w
@@ -352,7 +232,7 @@ def tilize_and_pack_bfloat16(data_2d: torch.Tensor, tile_h: int = 32, tile_w: in
             tiles[:, face_h:, face_w:].reshape(num_tiles, -1),
         ],
         axis=1,
-    )  # (N, 1024)
+    )
 
     float_bits = face_ordered.view(np.uint32)
     bf16_bits = (float_bits >> 16).astype(np.uint16)
@@ -360,22 +240,14 @@ def tilize_and_pack_bfloat16(data_2d: torch.Tensor, tile_h: int = 32, tile_w: in
 
 
 def tilize_and_pack_bfp4(data_2d: torch.Tensor, tile_h: int = 32, tile_w: int = 32) -> bytes:
-    """Tilize a 2-D tensor and pack as BFP4_b raw bytes.
-
-    Delegates to the C++ ``pack_as_bfp4_tiles`` via nanobind.
-    """
+    """Tilize a 2-D tensor and pack as BFP4_b raw bytes."""
     H, W = data_2d.shape
     data_np = data_2d.contiguous().float().numpy()
     return ttnn._ttnn.core.tilize_and_pack_bfp4_b(data_np, H, W, tile_h, tile_w)
 
 
 def pack_bfloat16_1x32(data: torch.Tensor) -> bytes:
-    """Pack a 1-row tensor as raw bfloat16 bytes with 1×32 tile layout.
-
-    For 1×32 tiles there is no face reordering; elements are stored
-    sequentially in tile-width chunks.  bfloat16 is the top 16 bits
-    of IEEE-754 float32.
-    """
+    """Pack a 1-row tensor as raw bfloat16 bytes with 1x32 tile layout."""
     flat = data.contiguous().float().reshape(-1).numpy()
     float_bits = flat.view(np.uint32)
     bf16_bits = (float_bits >> 16).astype(np.uint16)
