+++ models/demos/deepseek_v3_b1/weights/specs/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""DeepSeek-specific overlap and fusion group specifications."""
